Refactor: SizedProduct 엔티티와 Product 엔티티 병합[#52]
