+++ src/test/java/com/programmers/dev/kream/product/application/BrandServiceTest.java
@@ -3,7 +3,6 @@ package com.programmers.dev.kream.product.application;
 import com.programmers.dev.kream.product.domain.Brand;
 import com.programmers.dev.kream.product.domain.BrandRepository;
 import com.programmers.dev.kream.product.domain.ProductRepository;
-import com.programmers.dev.kream.product.domain.SizedProductRepository;
 import com.programmers.dev.kream.product.ui.dto.BrandResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
